dev-console-quiz-game-vanilla-JS
Quiz game within the developer console using only Vanilla JS. The project was hard, as some of the more advanced concepts covered in the video weren't in-depth enough to be able to understand application, so I had to turn to outside resources for further learning. Overall, good project to work on and I learned a lot.